moved helpers and documented
